Revert the last commit and put in the proper way to update the instrument selections.
Up until now in Jokosher the practise has been to update everything just in case. As can be seen in the previous commit, a nice and easy to understand solution to the instrument selection colour not updating is to update RecordingView whenever something on the instrument changes. However this is not good because the Update method in RecordingView removes all the InstrumentViewers, tells them all to update, and then repacks them all. This is overkill and degrades performance. 
The new way (which goes along with the new GObject system) is for only things that need to be updated to be updated. This commit makes only InstrumentViewer react to an instrument selection, and you will notice that it doesn't just call the Update() method either. It actually only executes the code that changes the instrument's colour so it looks selected, because this is the *only* thing that needs to be changed.
If something isn't updating (which is likely because of all the changes I've made with GObject), you should figure out which code needs to be called to make the change, and only run that code. Don't just call update on everything like we all did before.