Fix for bug #134181 (Mixdown actions don't display if user doesn't have Pidgin running).
We should incorporate a pre-loading hook so that extensions can check if the system satisfies their dependencies, and then Jokosher will load them or not, depending on this result.

This could lead to an improved extension manager window that informs users things like "Your extension was found, but it won't work properly because X,Y and Z are missing".